Protect against volumes in recycle bin
The Flocker functional tests expect only known volumes
to be present in the listing of volumes. Under some
cases it would be possible for volumes to be in the
recycle bin. These were not getting filtered out and
would end up being reported in the volume list.

Added filter to exclude any volumes in the recycle bin.